requirements.txt
tensorflow
keras
imutils
scipy
scikit-learn
pillow
aiortc==1.2.0
av==8.0.3
matplotlib==3.4.2
numpy==1.19.5
opencv_python==4.5.2.52
pydub==0.25.1
streamlit==0.81.1
streamlit_webrtc==0.20.0
typing_extensions==3.10.0.0